+++ tests/test_enum.rs
@@ -175,3 +175,18 @@ fn test_rename_enum_repr_correct() {
         py_assert!(py, var1, "repr(var1) == 'MyEnum.Variant'");
     })
 }
+
+#[pyclass]
+#[derive(Debug, PartialEq, Clone)]
+pub enum RenameVariantEnum {
+    #[pyo3(name = "VARIANT")]
+    Variant,
+}
+
+#[test]
+fn test_rename_variant_repr_correct() {
+    Python::with_gil(|py| {
+        let var1 = Py::new(py, RenameVariantEnum::Variant).unwrap();
+        py_assert!(py, var1, "repr(var1) == 'RenameVariantEnum.VARIANT'");
+    })
+}